Kind checker integrated into plugin
The kind checker now produces quite nice error messages for type synonyms
definitions.

The equality on kinds has been rewritten in the style of typeEq in order
to handle KiUnknown correctly.

Errors as lists of term-message pairs are only produced by kindOf
since this functions knows which type is to blame.

Comparision functions like kindEq or KindEqStar only produce a list
of messages that are attached to a type term by kindOf.