Proper type checking of imports
We write a signature line like

module-signature :
  module-id -> module-sig

to the desugarings of a module.

before type checking the definitions of a module, we build the
context induced by the imports by calling module-signature on imported
modules and translating the signature to a context (with qualified identifiers
if necessary).

Unfortunately, we type check a module twice: once to produce the signature line and once
for the error messages.